+++ covasim/sim.py
@@ -175,12 +175,8 @@ class Sim(cvb.BaseSim):
                 errormsg = f'You must supply one of n_days and end_day, not "{n_days}" and "{end_day}"'
                 raise ValueError(errormsg)
 
-        # Handle parameters specified by layer
-
-
-        # Try to figure out what the layer keys should be
+        # Layer keys -- first, try to figure out what the layer keys should be
         layer_keys = None # e.g. household, school
-        layer_pars = ['beta_layer', 'contacts', 'iso_factor', 'quar_factor']
         if self.people is not None:
             layer_keys = set(self.people.contacts.keys())
         elif isinstance(self['beta_layer'], dict):
@@ -188,13 +184,14 @@ class Sim(cvb.BaseSim):
         else:
             layer_keys = ['a'] # Assume this by default, corresponding to random/no layers
 
-        # Convert scalar layer parameters to dictionaries
+        # Layer keys -- convert scalar layer parameters to dictionaries
+        layer_pars = cvpar.layer_pars # The names of the parameters that are specified by layer
         for lp in layer_pars:
             val = self[lp]
             if sc.isnumber(val): # It's a scalar instead of a dict, assume it's all contacts
                 self[lp] = {k:val for k in layer_keys}
 
-        # Handle key mismaches
+        # Layer keys -- handle key mismaches
         for lp in layer_pars:
             lp_keys = set(self.pars[lp].keys())
             if not lp_keys == set(layer_keys):
@@ -203,13 +200,14 @@ class Sim(cvb.BaseSim):
                     errormsg += f'\n{lp2} = ' + ', '.join(self.pars[lp].keys())
                 raise sc.KeyNotFoundError(errormsg)
 
+        # Layer keys -- handle mismatches with the population
         if self.people is not None:
             pop_keys = set(self.people.contacts.keys())
             if pop_keys != layer_keys:
                 errormsg = f'Please update your parameter keys {layer_keys} to match population keys {pop_keys}. You may find sim.reset_layer_pars() helpful.'
                 raise sc.KeyNotFoundError(errormsg)
 
-        # Handle population data
+        # Moving on, handle population data
         popdata_choices = ['random', 'hybrid', 'clustered', 'synthpops']
         choice = self['pop_type']
         if choice not in popdata_choices:
@@ -254,7 +252,6 @@ class Sim(cvb.BaseSim):
         for key,label in cvd.result_stocks.items():
             self.results[f'n_{key}'] = init_res(label, color=dcols[key])
         self.results['n_susceptible'].scale = 'static'
-        self.results['bed_capacity']  = init_res('Bed demand relative to capacity', scale=False)
 
         # Other variables
         self.results['r_eff']         = init_res('Effective reproductive number', scale=False)
@@ -322,10 +319,7 @@ class Sim(cvb.BaseSim):
 
         # Create the seed infections
         inds = cvu.choose(self['pop_size'], self['pop_infected'])
-        self.people.infect(inds=inds)
-        for ind in inds:
-            self.people.transtree.linelist[ind] = dict(source=None, target=ind, date=self.t, layer='seed_infection')
-
+        self.people.infect(inds=inds, layer='seed_infection')
         return
 
 
@@ -371,17 +365,16 @@ class Sim(cvb.BaseSim):
         # Perform initial operations
         self.rescale() # Check if we need to rescale
         people   = self.people # Shorten this for later use
-        flows    = people.update_states_pre(t=t) # Update the state of everyone and count the flows
+        people.update_states_pre(t=t) # Update the state of everyone and count the flows
         contacts = people.update_contacts() # Compute new contacts
-        bed_max  = people.count('severe') > self['n_beds'] if self['n_beds'] else False # Check for a bed constraint
+        hosp_max = people.count('severe')   > self['n_beds_hosp'] if self['n_beds_hosp'] else False # Check for acute bed constraint
+        icu_max  = people.count('critical') > self['n_beds_icu']  if self['n_beds_icu']  else False # Check for ICU bed constraint
 
         # Randomly infect some people (imported infections)
         n_imports = cvu.poisson(self['n_imports']) # Imported cases
         if n_imports>0:
-            imporation_inds = cvu.choose(max_n=len(people), n=n_imports)
-            flows['new_infections'] += people.infect(inds=imporation_inds, bed_max=bed_max)
-            for ind in imporation_inds:
-                self.people.transtree.linelist[ind] = dict(source=None, target=ind, date=self.t, layer='importation')
+            importation_inds = cvu.choose(max_n=len(people), n=n_imports)
+            people.infect(inds=importation_inds, hosp_max=hosp_max, icu_max=icu_max, layer='importation')
 
         # Apply interventions
         for intervention in self['interventions']:
@@ -389,7 +382,7 @@ class Sim(cvb.BaseSim):
         if self['interv_func'] is not None: # Apply custom intervention function
             self['interv_func'](self)
 
-        flows = people.update_states_post(flows) # Check for state changes after interventions
+        people.update_states_post() # Check for state changes after interventions
 
         # Compute the probability of transmission
         beta         = cvd.default_float(self['beta'])
@@ -422,24 +415,16 @@ class Sim(cvb.BaseSim):
 
             # Calculate actual transmission
             for sources,targets in [[p1,p2], [p2,p1]]: # Loop over the contact network from p1->p2 and p2->p1
-                target_inds, edge_inds = cvu.compute_infections(beta, sources, targets, betas, rel_trans, rel_sus) # Calculate transmission!
-                flows['new_infections'] += people.infect(inds=target_inds, bed_max=bed_max) # Actually infect people
+                source_inds, target_inds = cvu.compute_infections(beta, sources, targets, betas, rel_trans, rel_sus) # Calculate transmission!
+                people.infect(inds=target_inds, hosp_max=hosp_max, icu_max=icu_max, source=source_inds, layer=lkey) # Actually infect people
 
-                # Store the transmission tree
-                for ind in edge_inds:
-                    source = sources[ind]
-                    target = targets[ind]
-                    transdict = dict(source=source, target=target, date=self.t, layer=lkey)
-                    self.people.transtree.linelist[target] = transdict
-                    self.people.transtree.targets[source].append(transdict)
 
         # Update counts for this time step: stocks
         for key in cvd.result_stocks.keys():
             self.results[f'n_{key}'][t] = people.count(key)
-        self.results['bed_capacity'][t] = self.results['n_severe'][t]/self['n_beds'] if self['n_beds'] else 0
 
         # Update counts for this time step: flows
-        for key,count in flows.items():
+        for key,count in people.flows.items():
             self.results[key][t] += count
 
         # Tidy up